fix: move FILE_STORAGES_TOKEN to IFileStorage.ts
