--- src/domain/storages/index.ts
@@ -1,1 +1,0 @@
-export const FILE_STORAGES_TOKEN = 'file_storages_token';
